cleanup download
