+++ Cargo.toml
@@ -9,6 +9,7 @@ members = [
 [workspace.package]
 edition = "2024"
 version = "2.0.0"
+license = "GPL-3.0"
 authors = ["C <claration@riseup.net>", "V <vendicated@riseup.net>"]
 repository = "https://github.com/Vencord/Installer"
