+++ crates/vencord-installer-cli/Cargo.toml
@@ -2,6 +2,7 @@
 name = "vencord-installer-cli"
 version.workspace = true
 edition.workspace = true
+license.workspace = true
 authors.workspace = true
 repository.workspace = true
