+++ crates/vencord-installer-core/Cargo.toml
@@ -2,6 +2,7 @@
 name = "vencord-installer-core"
 version.workspace = true
 edition.workspace = true
+license.workspace = true
 authors.workspace = true
 repository.workspace = true
 
@@ -13,6 +14,7 @@ serde_json = "1.0"
 serde = { version = "1.0", features = ["derive"] }
 regex = "1.12"
 thiserror = "2.0"
+dirs = "6.0.0"
 
 [target.'cfg(windows)'.dependencies]
 windows = { version = "0.62", features = [
@@ -24,6 +26,5 @@ windows = { version = "0.62", features = [
 ] }
 
 [features]
-default = ["openasar", "version_check"]
+default = ["openasar"]
 openasar = []
-version_check = []
