--- crates/vencord-installer-core/src/update/mod.rs
@@ -1,3 +1,0 @@
-pub mod download;
-#[cfg(feature = "version_check")]
-pub mod version_check;
