+++ crates/vencord-installer-gui/Cargo.toml
@@ -2,6 +2,7 @@
 name = "vencord-installer-gui"
 version.workspace = true
 edition.workspace = true
+license.workspace = true
 authors.workspace = true
 repository.workspace = true
