+++ crates/vencord-installer-gui/ui/globals.slint
@@ -21,7 +21,6 @@ export struct DiscordLocation {
     is_openasar: bool,
     is_flatpak: bool,
     is_scuffed: bool,
-    is_system_electron: bool,
 }
 
 export global DiscordLocationAdapter {
